ugnu: images: Add jetson image.
The jetson image is a base image using the jetson kernel and a custom
bootloader. The bootloader is currently not implemented and simply
inherits the extlinux bootloader, omitting the install step.

This is intended to get the building, even though it won't boot
yet. To build, use:

guix system -L ~/src/uguix image ~/src/uguix/ugnu/system/images/jetson.scm

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>