deb-download: Initial revision.
The url-fetch/deb is an <origin> method that will download a .deb
debian or ubuntu package and unpack it into the source tree. This is
mostly useful for nonfree packages that are distributed as debs
instead of source.

The rule hardcodes zstd compression for the data files (which is only
available in ubuntu's dpkg right now). More compression protocols will
be added with some additional customization in the future.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>